Action: Completed GetAccountInfoAsync(). Split MetatraderClient class across files.

+++ MetatraderSharp/MTsocketAPI.Responses/Account/Account.cs
@@ -46,28 +46,28 @@ public class Account
     public double Balance { get; set; }
 
     [JsonProperty("CREDIT")]
-    public int Credit { get; set; }
+    public double Credit { get; set; }
 
     [JsonProperty("PROFIT")]
-    public int Profit { get; set; }
+    public double Profit { get; set; }
 
     [JsonProperty("EQUITY")]
     public double Equity { get; set; }
 
     [JsonProperty("MARGIN")]
-    public int Margin { get; set; }
+    public double Margin { get; set; }
 
     [JsonProperty("MARGIN_FREE")]
     public double MarginFree { get; set; }
 
     [JsonProperty("MARGIN_LEVEL")]
-    public int MarginLevel { get; set; }
+    public double MarginLevel { get; set; }
 
     [JsonProperty("MARGIN_SO_CAL")]
-    public int MarginSoCal { get; set; }
+    public double MarginSoCal { get; set; }
 
     [JsonProperty("MARGIN_SO_SO")]
-    public int MarginSoSo { get; set; }
+    public double MarginSoSo { get; set; }
 
     [JsonProperty("ERROR_ID")]
     public int ErrorID { get; set; }
